pyproject.toml
[tool.poetry]
name = "anonymizer"
version = "17.2.1rc7"
description = "RSNA DICOM Anonymizer"
authors = ["Algorhythm Software <michael@dx.life>"]
license = "RSNA-MIRC Public License"
readme = "README.md"
homepage = "https://github.com/RSNA/anonymizer"
packages = [{ include = "anonymizer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.7"
numpy = "<2.0.0"
customtkinter = "*"
pillow = "*"
tkhtmlview = "*"
pydicom = ">=2.4,<2.5"
pynetdicom = "*"
ifaddr = "*"
boto3 = "*"
pywin32-ctypes = "*"
pefile = "*"
openpyxl = "*"
psutil = "*"
pylibjpeg = {extras = ["all"], version = "*"}
opencv-python-headless = "*"
easyocr = "*"
cryptography = "*"
requests = "*"
requests-oauthlib = "*"
poetry = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "*"
black = "*"
pytest = "*"
pytest-cov = "^6.0.0"
python-dotenv = "*"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C90"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/anonymizer --cov-report=html"

[tool.poetry.scripts]
test = "pytest:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


